Added vpn_required mark to ptest.ini Added vpn_required mark to any tests that it applies to Moved cognito login page into the login directory
