+++ pytest.ini
@@ -32,3 +32,4 @@ markers =
     compartment1: only for compartment 1
     compartment2: only for compartment 2
     compartment3: only for compartment 3
+    vpn_required: for tests that require a VPN connection
